requirements.txt
feedparser>=6.0.0
pyyaml>=6.0
# Optional — only needed if you want AI-generated summaries
anthropic>=0.40.0
openai>=1.50.0
